requirements.txt
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
isodate==0.6.1
msrest==0.6.21
oauthlib==3.2.0
Pillow==9.0.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
urllib3==1.26.9
